Strange Counter
Bob has a strange counter. At the first second, , it displays the number . At each subsequent second, the number displayed by the counter decrements by .

The counter counts down in cycles. In the second after the counter counts down to , the number becomes the initial number for that countdown cycle and then continues counting down from the new initial number in a new cycle. The diagram below shows the counter values for each time in the first three cycles:

[strange(1).png]

Given a time, , find and print the value displayed by the counter at time .

Input Format

A single integer denoting the value of .

Constraints

Subtask

    for of the maximum score.

Output Format

Print the value displayed by the strange counter at the given time .

Sample Input

4

Sample Output

6

Explanation

Time marks the beginning of the second cycle in which the counter displays a number that is double the initial number displayed at the beginning of the previous cycle (i.e., ). This is also shown in the diagram in the Problem Statement above.